requirements.txt
asgiref==3.4.1
autopep8==1.5.7
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.1.13
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Naked==0.1.31
psycopg2==2.8.6
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.3.0
python-decouple==3.5
python-http-client==3.3.3
pytz==2021.3
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
